+++ src/main/java/lotto/constants/LottoConstant.java
@@ -1,4 +1,4 @@
-package lotto.view.constants;
+package lotto.constants;
 
 public class LottoConstant {
     public static final int LOTTO_PRICE = 1000;
