+++ src/test/java/lotto/view/model/MoneyTest.java
@@ -1,11 +1,12 @@
 package lotto.view.model;
 
+import lotto.model.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import static lotto.view.constants.LottoConstant.LOTTO_PRICE;
+import static lotto.constants.LottoConstant.LOTTO_PRICE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
