+++ src/test/java/lotto/view/utils/InputUtilsTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.params.provider.CsvSource;
 
 import java.util.List;
 
-import static lotto.view.utils.InputUtils.*;
+import static lotto.utils.InputUtils.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class InputUtilsTest {
